Rename helpers folder, move some vue filters
