+++ client/js/helpers/colorClass.js
