+++ client/js/helpers/friendlysize.js
