+++ client/js/helpers/ircmessageparser/anyIntersection.js
