+++ client/js/helpers/ircmessageparser/cleanIrcMessage.js
