+++ client/js/helpers/ircmessageparser/findChannels.js
