+++ client/js/helpers/ircmessageparser/findEmoji.js
