+++ client/js/helpers/ircmessageparser/findLinks.js
