+++ client/js/helpers/ircmessageparser/findNames.js
