+++ client/js/helpers/ircmessageparser/merge.js
