+++ client/js/helpers/ircmessageparser/parseStyle.js
