+++ client/js/helpers/simplemap.json
