+++ src/plugins/irc-events/link.js
@@ -5,8 +5,8 @@ const got = require("got");
 const URL = require("url").URL;
 const mime = require("mime-types");
 const Helper = require("../../helper");
-const cleanIrcMessage = require("../../../client/js/libs/handlebars/ircmessageparser/cleanIrcMessage");
-const findLinks = require("../../../client/js/libs/handlebars/ircmessageparser/findLinks");
+const cleanIrcMessage = require("../../../client/js/helpers/ircmessageparser/cleanIrcMessage");
+const findLinks = require("../../../client/js/helpers/ircmessageparser/findLinks");
 const storage = require("../storage");
 const currentFetchPromises = new Map();
 const imageTypeRegex = /^image\/.+/;
