+++ test/client/js/helpers/ircmessageparser/findEmoji.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const expect = require("chai").expect;
-const findEmoji = require("../../../../../../client/js/libs/handlebars/ircmessageparser/findEmoji");
+const findEmoji = require("../../../../../client/js/helpers/ircmessageparser/findEmoji");
 
 describe("findEmoji", () => {
 	it("should find default emoji presentation character", () => {
